+++ src/main/java/nl/dtls/fairdatapoint/service/reset/FactoryDefaults.java
@@ -34,28 +34,21 @@ import nl.dtls.fairdatapoint.entity.shape.ShapeType;
 import nl.dtls.fairdatapoint.entity.user.User;
 import nl.dtls.fairdatapoint.entity.user.UserRole;
 import nl.dtls.fairdatapoint.service.shape.ShapeShaclUtils;
+import nl.dtls.fairdatapoint.util.KnownUUIDs;
 import nl.dtls.fairdatapoint.vocabulary.DATACITE;
 import nl.dtls.fairdatapoint.vocabulary.FDP;
 import nl.dtls.fairdatapoint.vocabulary.R3D;
-import nl.dtls.fairdatapoint.vocabulary.Sio;
-import org.apache.commons.codec.digest.DigestUtils;
 import org.bson.BasicBSONObject;
 import org.bson.Document;
 import org.bson.types.BasicBSONList;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Statement;
-import org.eclipse.rdf4j.model.vocabulary.DCTERMS;
-import org.eclipse.rdf4j.model.vocabulary.FOAF;
-import org.eclipse.rdf4j.model.vocabulary.RDF;
-import org.eclipse.rdf4j.model.vocabulary.RDFS;
+import org.eclipse.rdf4j.model.vocabulary.*;
 
-import java.io.IOException;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
-import static java.lang.String.format;
 import static nl.dtls.fairdatapoint.util.ValueFactoryHelper.*;
 
 public class FactoryDefaults {
@@ -63,7 +56,7 @@ public class FactoryDefaults {
     //== USERS
     // Changes: Migration_0001_Init
     public static final User USER_ALBERT = User.builder()
-            .uuid("7e64818d-6276-46fb-8bb1-732e6e09f7e9")
+            .uuid(KnownUUIDs.USER_ALBERT_UUID)
             .firstName("Albert")
             .lastName("Einstein")
             .email("albert.einstein@example.com")
@@ -72,7 +65,7 @@ public class FactoryDefaults {
             .build();
 
     public static final User USER_NIKOLA = User.builder()
-            .uuid("b5b92c69-5ed9-4054-954d-0121c29b6800")
+            .uuid(KnownUUIDs.USER_NIKOLA_UUID)
             .firstName("Nikola")
             .lastName("Tesla")
             .email("nikola.tesla@example.com")
@@ -83,7 +76,7 @@ public class FactoryDefaults {
     //== MEMBERSHIPS
     // Changes: Migration_0001_Init, Migration_0004_ResourceDefinition
     public static final Membership MEMBERSHIP_OWNER = Membership.builder()
-            .uuid("49f2bcfd-ef0a-4a3a-a1a3-0fc72a6892a8")
+            .uuid(KnownUUIDs.MEMBERSHIP_OWNER_UUID)
             .name("Owner")
             .permissions(List.of(
                     new MembershipPermission(2, 'W'),
@@ -92,52 +85,43 @@ public class FactoryDefaults {
                     new MembershipPermission(16, 'A')
             ))
             .allowedEntities(List.of(
-                    "a0949e72-4466-4d53-8900-9436d1049a4b",
-                    "2f08228e-1789-40f8-84cd-28e3288c3604",
-                    "02c649de-c579-43bb-b470-306abdc808c7"
+                    KnownUUIDs.RD_CATALOG_UUID,
+                    KnownUUIDs.RD_DATASET_UUID,
+                    KnownUUIDs.RD_DISTRIBUTION_UUID
             ))
             .build();
 
     public static final Membership MEMBERSHIP_DATA_PROVIDER = Membership.builder()
-            .uuid("87a2d984-7db2-43f6-805c-6b0040afead5")
+            .uuid(KnownUUIDs.MEMBERSHIP_DATAPROVIDER_UUID)
             .name("Data Provider")
             .permissions(List.of(
                     new MembershipPermission(4, 'C')
             ))
             .allowedEntities(List.of(
-                    "a0949e72-4466-4d53-8900-9436d1049a4b"
+                    KnownUUIDs.RD_CATALOG_UUID
             ))
             .build();
 
-    //== SHAPE UUIDS
-    private static final String SHAPE_RESOURCE_UUID = "6a668323-3936-4b53-8380-a4fd2ed082ee";
-
-    private static final String SHAPE_REPOSITORY_UUID = "a92958ab-a414-47e6-8e17-68ba96ba3a2b";
-
-    private static final String SHAPE_CATALOG_UUID = "2aa7ba63-d27a-4c0e-bfa6-3a4e250f4660";
-
-    private static final String SHAPE_DATASET_UUID = "866d7fb8-5982-4215-9c7c-18d0ed1bd5f3";
-
-    private static final String SHAPE_DISTRIBUTION_UUID = "ebacbf83-cd4f-4113-8738-d73c0735b0ab";
-
     //== RESOURCE DEFINITIONS
-    // Changes: Migration_0002_CustomMetamodel, Migration_0004_ResourceDefinition
-    public static final ResourceDefinition RESOURCE_DEFINITION_REPOSITORY = ResourceDefinition.builder()
-            .uuid("77aaad6a-0136-4c6e-88b9-07ffccd0ee4c")
-            .name("Repository")
+    // Changes: Migration_0002_CustomMetamodel, Migration_0004_ResourceDefinition, Migration_0010_ComplyFDPO
+    public static final ResourceDefinition RESOURCE_DEFINITION_FDP = ResourceDefinition.builder()
+            .uuid(KnownUUIDs.RD_FDP_UUID)
+            .name("FAIR Data Point")
             .urlPrefix("")
             .shapeUuids(List.of(
-                    SHAPE_RESOURCE_UUID,
-                    SHAPE_REPOSITORY_UUID
+                    KnownUUIDs.SHAPE_RESOURCE_UUID,
+                    KnownUUIDs.SHAPE_DATASERVICE_UUID,
+                    KnownUUIDs.SHAPE_METADATASERVICE_UUID,
+                    KnownUUIDs.SHAPE_FDP_UUID
             ))
             .children(List.of(
                     ResourceDefinitionChild.builder()
-                            .resourceDefinitionUuid("a0949e72-4466-4d53-8900-9436d1049a4b")
-                            .relationUri("http://www.re3data.org/schema/3-0#dataCatalog")
+                            .resourceDefinitionUuid(KnownUUIDs.RD_CATALOG_UUID)
+                            .relationUri(FDP.METADATACATALOG.stringValue())
                             .listView(
                                     ResourceDefinitionChildListView.builder()
                                             .title("Catalogs")
-                                            .tagsUri("http://www.w3.org/ns/dcat#themeTaxonomy")
+                                            .tagsUri(DCAT.THEME_TAXONOMY.stringValue())
                                             .metadata(List.of())
                                             .build()
                             )
@@ -147,21 +131,21 @@ public class FactoryDefaults {
             .build();
 
     public static final ResourceDefinition RESOURCE_DEFINITION_CATALOG = ResourceDefinition.builder()
-            .uuid("a0949e72-4466-4d53-8900-9436d1049a4b")
+            .uuid(KnownUUIDs.RD_CATALOG_UUID)
             .name("Catalog")
             .urlPrefix("catalog")
             .shapeUuids(List.of(
-                    SHAPE_RESOURCE_UUID,
-                    SHAPE_CATALOG_UUID
+                    KnownUUIDs.SHAPE_RESOURCE_UUID,
+                    KnownUUIDs.SHAPE_CATALOG_UUID
             ))
             .children(List.of(
                     ResourceDefinitionChild.builder()
-                            .resourceDefinitionUuid("2f08228e-1789-40f8-84cd-28e3288c3604")
-                            .relationUri("http://www.w3.org/ns/dcat#dataset")
+                            .resourceDefinitionUuid(KnownUUIDs.RD_DATASET_UUID)
+                            .relationUri(DCAT.HAS_DATASET.stringValue())
                             .listView(
                                     ResourceDefinitionChildListView.builder()
                                             .title("Datasets")
-                                            .tagsUri("http://www.w3.org/ns/dcat#theme")
+                                            .tagsUri(DCAT.THEME.stringValue())
                                             .metadata(List.of())
                                             .build()
                             )
@@ -171,17 +155,17 @@ public class FactoryDefaults {
             .build();
 
     public static final ResourceDefinition RESOURCE_DEFINITION_DATASET = ResourceDefinition.builder()
-            .uuid("2f08228e-1789-40f8-84cd-28e3288c3604")
+            .uuid(KnownUUIDs.RD_DATASET_UUID)
             .name("Dataset")
             .urlPrefix("dataset")
             .shapeUuids(List.of(
-                    SHAPE_RESOURCE_UUID,
-                    SHAPE_DATASET_UUID
+                    KnownUUIDs.SHAPE_RESOURCE_UUID,
+                    KnownUUIDs.SHAPE_DATASET_UUID
             ))
             .children(List.of(
                     ResourceDefinitionChild.builder()
-                            .resourceDefinitionUuid("02c649de-c579-43bb-b470-306abdc808c7")
-                            .relationUri("http://www.w3.org/ns/dcat#distribution")
+                            .resourceDefinitionUuid(KnownUUIDs.RD_DISTRIBUTION_UUID)
+                            .relationUri(DCAT.HAS_DISTRIBUTION.stringValue())
                             .listView(
                                     ResourceDefinitionChildListView.builder()
                                             .title("Distributions")
@@ -189,7 +173,7 @@ public class FactoryDefaults {
                                             .metadata(List.of(
                                                     new ResourceDefinitionChildListViewMetadata(
                                                             "Media Type",
-                                                            "http://www.w3.org/ns/dcat#mediaType"
+                                                            DCAT.MEDIA_TYPE.stringValue()
                                                     )
                                             ))
                                             .build()
@@ -200,32 +184,32 @@ public class FactoryDefaults {
             .build();
 
     public static ResourceDefinition RESOURCE_DEFINITION_DISTRIBUTION = ResourceDefinition.builder()
-            .uuid("02c649de-c579-43bb-b470-306abdc808c7")
+            .uuid(KnownUUIDs.RD_DISTRIBUTION_UUID)
             .name("Distribution")
             .urlPrefix("distribution")
             .shapeUuids(List.of(
-                    SHAPE_RESOURCE_UUID,
-                    SHAPE_DISTRIBUTION_UUID
+                    KnownUUIDs.SHAPE_RESOURCE_UUID,
+                    KnownUUIDs.SHAPE_DISTRIBUTION_UUID
             ))
             .children(List.of())
             .externalLinks(List.of(
                     new ResourceDefinitionLink(
                             "Access online",
-                            "http://www.w3.org/ns/dcat#accessURL"
+                            DCAT.ACCESS_URL.stringValue()
                     ),
                     new ResourceDefinitionLink(
                             "Download",
-                            "http://www.w3.org/ns/dcat#downloadURL"
+                            DCAT.DOWNLOAD_URL.stringValue()
                     )
             ))
             .build();
 
     //== SHAPES
-    //== Changes: Migration_0003_ShapeDefinition, Migration_0005_UpdateShapeDefinition, Migration_0006_ShapesSharing
+    //== Changes: Migration_0003_ShapeDefinition, Migration_0005_UpdateShapeDefinition, Migration_0006_ShapesSharing, Migration_0010_ComplyFDPO
     public static Shape shapeResource() throws Exception {
         String definition = loadShape("shape-resource.ttl");
         return Shape.builder()
-                .uuid(SHAPE_RESOURCE_UUID)
+                .uuid(KnownUUIDs.SHAPE_RESOURCE_UUID)
                 .name("Resource")
                 .type(ShapeType.INTERNAL)
                 .published(false)
@@ -234,11 +218,35 @@ public class FactoryDefaults {
                 .build();
     }
 
-    public static Shape shapeRepository() throws Exception {
-        String definition = loadShape("shape-repository.ttl");
+    public static Shape shapeFDP() throws Exception {
+        String definition = loadShape("shape-fdp.ttl");
         return Shape.builder()
-                .uuid(SHAPE_REPOSITORY_UUID)
-                .name("Repository")
+                .uuid(KnownUUIDs.SHAPE_FDP_UUID)
+                .name("FAIR Data Point")
+                .type(ShapeType.INTERNAL)
+                .published(false)
+                .definition(definition)
+                .targetClasses(ShapeShaclUtils.extractTargetClasses(definition))
+                .build();
+    }
+
+    public static Shape shapeDataService() throws Exception {
+        String definition = loadShape("shape-data-service.ttl");
+        return Shape.builder()
+                .uuid(KnownUUIDs.SHAPE_DATASERVICE_UUID)
+                .name("Data Service")
+                .type(ShapeType.INTERNAL)
+                .published(false)
+                .definition(definition)
+                .targetClasses(ShapeShaclUtils.extractTargetClasses(definition))
+                .build();
+    }
+
+    public static Shape shapeMetadataService() throws Exception {
+        String definition = loadShape("shape-metadata-service.ttl");
+        return Shape.builder()
+                .uuid(KnownUUIDs.SHAPE_METADATASERVICE_UUID)
+                .name("Metadata Service")
                 .type(ShapeType.INTERNAL)
                 .published(false)
                 .definition(definition)
@@ -249,7 +257,7 @@ public class FactoryDefaults {
     public static Shape shapeCatalog() throws Exception {
         String definition = loadShape("shape-catalog.ttl");
         return Shape.builder()
-                .uuid(SHAPE_CATALOG_UUID)
+                .uuid(KnownUUIDs.SHAPE_CATALOG_UUID)
                 .name("Catalog")
                 .type(ShapeType.INTERNAL)
                 .published(false)
@@ -261,7 +269,7 @@ public class FactoryDefaults {
     public static Shape shapeDataset() throws Exception {
         String definition = loadShape("shape-dataset.ttl");
         return Shape.builder()
-                .uuid(SHAPE_DATASET_UUID)
+                .uuid(KnownUUIDs.SHAPE_DATASET_UUID)
                 .name("Dataset")
                 .type(ShapeType.CUSTOM)
                 .published(false)
@@ -273,7 +281,7 @@ public class FactoryDefaults {
     public static Shape shapeDistribution() throws Exception {
         String definition = loadShape("shape-distribution.ttl");
         return Shape.builder()
-                .uuid(SHAPE_DISTRIBUTION_UUID)
+                .uuid(KnownUUIDs.SHAPE_DISTRIBUTION_UUID)
                 .name("Distribution")
                 .type(ShapeType.CUSTOM)
                 .published(false)
@@ -358,6 +366,39 @@ public class FactoryDefaults {
         FactoryDefaults.add(s, identifierIri, DCTERMS.IDENTIFIER, l(persistentUrl), baseUrl);
         // Repository Identifier
         FactoryDefaults.add(s, R3D.REPOSITORYIDENTIFIER, identifierIri, baseUrl);
+        // Access Rights
+        IRI arIri = i(persistentUrl + "#accessRights");
+        FactoryDefaults.add(s, DCTERMS.ACCESS_RIGHTS, arIri, baseUrl);
+        FactoryDefaults.add(s, arIri, RDF.TYPE, DCTERMS.RIGHTS_STATEMENT, baseUrl);
+        FactoryDefaults.add(s, arIri, DCTERMS.DESCRIPTION, l(accessRightsDescription), baseUrl);
+        // Publisher
+        IRI publisherIri = i(persistentUrl + "#publisher");
+        FactoryDefaults.add(s, DCTERMS.PUBLISHER, publisherIri, baseUrl);
+        FactoryDefaults.add(s, publisherIri, RDF.TYPE, FOAF.AGENT, baseUrl);
+        FactoryDefaults.add(s, publisherIri, FOAF.NAME, l("Default Publisher"), baseUrl);
+        return s;
+    }
+
+    public static List<Statement> fdpStatements(String persistentUrl, IRI license, IRI language, String accessRightsDescription) {
+        List<Statement> s = new ArrayList<>();
+        IRI baseUrl = i(persistentUrl);
+        FactoryDefaults.add(s, RDF.TYPE, FDP.FAIRDATAPOINT, baseUrl);
+        FactoryDefaults.add(s, RDF.TYPE, FDP.METADATASERVICE, baseUrl);
+        FactoryDefaults.add(s, RDF.TYPE, DCAT.DATA_SERVICE, baseUrl);
+        FactoryDefaults.add(s, RDF.TYPE, DCAT.RESOURCE, baseUrl);
+        FactoryDefaults.add(s, DCTERMS.TITLE, l("My FAIR Data Point"), baseUrl);
+        FactoryDefaults.add(s, RDFS.LABEL, l("My FAIR Data Point"), baseUrl);
+        FactoryDefaults.add(s, DCTERMS.HAS_VERSION, l(1.0f), baseUrl);
+        FactoryDefaults.add(s, FDP.METADATAISSUED, l(OffsetDateTime.now()), baseUrl);
+        FactoryDefaults.add(s, FDP.METADATAMODIFIED, l(OffsetDateTime.now()), baseUrl);
+        FactoryDefaults.add(s, DCTERMS.LICENSE, license, baseUrl);
+        FactoryDefaults.add(s, DCTERMS.DESCRIPTION, l(LIPSUM_TEXT), baseUrl);
+        FactoryDefaults.add(s, DCTERMS.LANGUAGE, language, baseUrl);
+        // Identifier
+        IRI identifierIri = i(persistentUrl + "#identifier");
+        FactoryDefaults.add(s, FDP.METADATAIDENTIFIER, identifierIri, baseUrl);
+        FactoryDefaults.add(s, identifierIri, RDF.TYPE, DATACITE.IDENTIFIER, baseUrl);
+        FactoryDefaults.add(s, identifierIri, DCTERMS.IDENTIFIER, l(persistentUrl), baseUrl);
         // Access Rights
         IRI arIri = i(persistentUrl + "#accessRights");
         FactoryDefaults.add(s, DCTERMS.ACCESS_RIGHTS, arIri, baseUrl);